// engine/src/types/fromApi.ts
  export const CREATE_ORDER = "CREATE_ORDER";
  export const CANCEL_ORDER = "CANCEL_ORDER";
  export const ON_RAMP = "ON_RAMP";
  export const GET_DEPTH = "GET_DEPTH";
  export const GET_OPEN_ORDERS = "GET_OPEN_ORDERS";

  //TODO: Can we share the types between the api and the engine?
  export type MessageFromApi =
    | {
        type: typeof CREATE_ORDER;
        data: {
          market: string;
          price: string;
          quantity: string;
          side: "buy" | "sell";
          userId: string;
        };
      }
    | {
        type: typeof CANCEL_ORDER;
        data: {
          orderId: string;
          market: string;
        };
      }
    | {
        type: typeof ON_RAMP;
        data: {
          amount: string;
          userId: string;
          txnId: string;
        };
      }
    | {
        type: typeof GET_DEPTH;
        data: {
          market: string;
        };
      }
    | {
        type: typeof GET_OPEN_ORDERS;
        data: {
          userId: string;
          market: string;
        };
      };